Move config functions from utils.py to config.py
