+++ multiscanner/common/dir_monitor.py
@@ -21,7 +21,7 @@ from watchdog.observers import Observer
 
 from multiscanner import CONFIG as MS_CONFIG
 from multiscanner import multiscan, parse_reports
-from multiscanner.common import utils
+from multiscanner.config import get_config_path
 from multiscanner.storage import storage
 
 logger = logging.getLogger(__name__)
@@ -81,7 +81,7 @@ def start_observer(directory, work_queue, recursive=False):
 def multiscanner_process(work_queue, config, batch_size, wait_seconds, delete, exit_signal):
     filelist = []
     time_stamp = None
-    storage_conf = utils.get_config_path(config, 'storage')
+    storage_conf = get_config_path(config, 'storage')
     storage_handler = storage.StorageHandler(configfile=storage_conf)
     while not exit_signal.value:
         time.sleep(1)
